Don’t use reset and normalize css anymore!
Use a bare bones reset + normalize css instead. It is named `foundation` and it only includes the minimal required (opinionated) resets and normalizations! The old reset and normalize css files are still in the repository in case some extra reset/normalization has to be done… (re)add those the the foundation css on a project basis!