Pass on key-value filter pair with cookies (instead of in the url).
It is cleaner, less comples, and if a filtered blog post url is shared, the key-value filter pair is not part of it and dose noet influence the receivers browsing flow. If no cookies are allowd than the pagination between filtered posts will not work, but that is perfectly fine, and no deal breaker in experience.